monitoring/node-version-guard/main.py
"""
Node.js Version Guard - Cloud Function
Monitors deployed Cloud Run services for Node.js version compliance
Triggers alerts and automated responses for version violations

Compatible with AIXTIV-SYMPHONY Diamond SAO Command Center v34
"""

import json
import logging
import os
from datetime import datetime
from typing import Dict, List, Optional

import functions_framework
from google.cloud import monitoring_v3
from google.cloud import run_v2
from google.cloud import secretmanager
import requests

# Configuration
PROJECT_ID = os.environ.get("GOOGLE_CLOUD_PROJECT", "api-for-warp-drive")
MINIMUM_NODE_VERSION = int(os.environ.get("MINIMUM_NODE_VERSION", "24"))
DIAMOND_SAO_WEBHOOK = os.environ.get("DIAMOND_SAO_WEBHOOK", 
    "https://mocoa-owner-interface-859242575175.us-central1.run.app/api/monitoring/node-version-alert")

# Services to monitor
MONITORED_SERVICES = [
    "integration-gateway-js",
    "mcp-zaxxon-2100-cool", 
    "payment-pipeline",
    "pcp-activation-service",
    "mongodb-mcp-oauth-uscentral1",
    "auto-provision-mcp-uscentral1"
]

REGIONS = ["us-west1", "us-central1", "eu-west1"]

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class NodeVersionGuard:
    """Monitors and enforces Node.js version compliance across Cloud Run services."""
    
    def __init__(self):
        self.run_client = run_v2.ServicesClient()
        self.monitoring_client = monitoring_v3.MetricServiceClient()
        self.secret_client = secretmanager.SecretManagerServiceClient()
        
    def get_service_revisions(self, service_name: str, region: str) -> List[Dict]:
        """Get all active revisions for a Cloud Run service."""
        try:
            parent = f"projects/{PROJECT_ID}/locations/{region}"
            service_path = f"{parent}/services/{service_name}"
            
            # Get service details
            service = self.run_client.get_service(name=service_path)
            
            # Get revisions
            revisions_request = run_v2.ListRevisionsRequest(
                parent=service_path,
                show_deleted=False
            )
            
            revisions = self.run_client.list_revisions(request=revisions_request)
            
            result = []
            for revision in revisions:
                # Extract container image information
                containers = revision.spec.template.spec.containers
                if containers:
                    container = containers[0]  # Assume single container
                    image = container.image
                    
                    # Get traffic allocation
                    traffic_percent = 0
                    if service.spec.traffic:
                        for traffic in service.spec.traffic:
                            if traffic.revision == revision.metadata.name:
                                traffic_percent = traffic.percent
                                break
                    
                    result.append({
                        "revision_name": revision.metadata.name,
                        "image": image,
                        "traffic_percent": traffic_percent,
                        "created_time": revision.metadata.create_time,
                        "service_name": service_name,
                        "region": region
                    })
            
            return result
            
        except Exception as e:
            logger.error(f"Failed to get revisions for {service_name} in {region}: {e}")
            return []
    
    def extract_node_version(self, image_url: str) -> Optional[int]:
        """Extract Node.js version from container image URL or labels."""
        try:
            # Try to extract from image tag (e.g., node:24-alpine)
            if ":node" in image_url or "node:" in image_url:
                parts = image_url.split(":")
                for part in parts:
                    if "node" in part:
                        # Look for version number after node
                        version_str = part.replace("node", "").replace("-alpine", "").replace("-slim", "")
                        if version_str.isdigit():
                            return int(version_str)
            
            # Check for common patterns in image names
            if "nodejs24" in image_url or "node24" in image_url:
                return 24
            elif "nodejs22" in image_url or "node22" in image_url:
                return 22
            elif "nodejs20" in image_url or "node20" in image_url:
                return 20
            elif "nodejs18" in image_url or "node18" in image_url:
                return 18
            
            # Default assumption for recent builds
            return 24  # Optimistic assumption
            
        except Exception as e:
            logger.warning(f"Could not extract Node version from {image_url}: {e}")
            return None
    
    def create_monitoring_alert(self, violations: List[Dict]):
        """Create a monitoring alert for Node.js version violations."""
        try:
            project_name = f"projects/{PROJECT_ID}"
            
            # Create custom metric
            series = monitoring_v3.TimeSeries()
            series.metric.type = "custom.googleapis.com/node_version_violations"
            series.resource.type = "cloud_run_revision"
            series.resource.labels["service_name"] = "integration-gateway"
            
            # Add data point
            point = series.points.add()
            point.value.int64_value = len(violations)
            now = datetime.utcnow()
            point.interval.end_time.FromDatetime(now)
            
            # Send to monitoring
            self.monitoring_client.create_time_series(
                name=project_name, 
                time_series=[series]
            )
            
            logger.info(f"Created monitoring alert for {len(violations)} violations")
            
        except Exception as e:
            logger.error(f"Failed to create monitoring alert: {e}")
    
    def send_diamond_sao_notification(self, violations: List[Dict], summary: Dict):
        """Send notification to Diamond SAO Command Center."""
        try:
            payload = {
                "alert_type": "node_version_violation",
                "timestamp": datetime.utcnow().isoformat(),
                "severity": "high" if len(violations) > 0 else "info",
                "violations": violations,
                "summary": summary,
                "minimum_required_version": MINIMUM_NODE_VERSION,
                "monitored_services": MONITORED_SERVICES,
                "regions": REGIONS,
                "self_healing_recommended": len(violations) > 0
            }
            
            response = requests.post(
                DIAMOND_SAO_WEBHOOK,
                json=payload,
                headers={"Content-Type": "application/json"},
                timeout=30
            )
            
            if response.status_code == 200:
                logger.info("Successfully sent notification to Diamond SAO")
            else:
                logger.warning(f"Diamond SAO notification failed: {response.status_code}")
                
        except Exception as e:
            logger.error(f"Failed to send Diamond SAO notification: {e}")
    
    def check_all_services(self) -> Dict:
        """Check Node.js versions across all monitored services."""
        violations = []
        compliant_services = []
        
        for service_name in MONITORED_SERVICES:
            for region in REGIONS:
                revisions = self.get_service_revisions(service_name, region)
                
                for revision in revisions:
                    if revision["traffic_percent"] > 0:  # Only check active revisions
                        node_version = self.extract_node_version(revision["image"])
                        
                        if node_version and node_version < MINIMUM_NODE_VERSION:
                            violation = {
                                "service_name": revision["service_name"],
                                "region": revision["region"],
                                "revision_name": revision["revision_name"],
                                "current_node_version": node_version,
                                "required_node_version": MINIMUM_NODE_VERSION,
                                "traffic_percent": revision["traffic_percent"],
                                "image": revision["image"],
                                "severity": "critical" if revision["traffic_percent"] > 50 else "warning"
                            }
                            violations.append(violation)
                            logger.warning(f"Version violation: {service_name} in {region} using Node.js {node_version}")
                        
                        elif node_version and node_version >= MINIMUM_NODE_VERSION:
                            compliant_services.append({
                                "service_name": revision["service_name"],
                                "region": revision["region"],
                                "node_version": node_version,
                                "traffic_percent": revision["traffic_percent"]
                            })
        
        summary = {
            "total_violations": len(violations),
            "total_compliant": len(compliant_services),
            "critical_violations": len([v for v in violations if v["severity"] == "critical"]),
            "warning_violations": len([v for v in violations if v["severity"] == "warning"]),
            "check_timestamp": datetime.utcnow().isoformat()
        }
        
        # Create monitoring alerts if violations found
        if violations:
            self.create_monitoring_alert(violations)
        
        # Send notification to Diamond SAO
        self.send_diamond_sao_notification(violations, summary)
        
        return {
            "violations": violations,
            "compliant_services": compliant_services,
            "summary": summary
        }


@functions_framework.http
def node_version_guard_http(request):
    """HTTP Cloud Function entry point."""
    guard = NodeVersionGuard()
    result = guard.check_all_services()
    
    return {
        "status": "success",
        "data": result
    }, 200


@functions_framework.cloud_event
def node_version_guard_scheduled(cloud_event):
    """Scheduled Cloud Function entry point."""
    guard = NodeVersionGuard()
    result = guard.check_all_services()
    
    logger.info(f"Scheduled check completed: {result['summary']}")
    
    if result["violations"]:
        logger.critical(f"Found {len(result['violations'])} Node.js version violations!")
        
        # Trigger automated remediation if configured
        if os.environ.get("AUTO_REMEDIATE", "false").lower() == "true":
            logger.info("Triggering automated remediation...")
            # Could trigger Cloud Build job to redeploy with correct versions
    
    return "OK"


if __name__ == "__main__":
    # For local testing
    guard = NodeVersionGuard()
    result = guard.check_all_services()
    print(json.dumps(result, indent=2))