monitoring/node-version-guard/requirements.txt
functions-framework==3.*
google-cloud-monitoring==2.*
google-cloud-run==0.*
google-cloud-secret-manager==2.*
requests==2.*